requirements.txt
python-telegram-bot[ext]==21.10
pytesseract==0.3.13
Pillow==11.1.0
openai==1.61.0
openpyxl==3.1.5
pydantic==2.10.6
python-dotenv==1.0.1
